fix(block): Runtime db error and populate total output in block table and data type changes

+++ components/stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/domain/Epoch.java
@@ -8,7 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Date;
+import java.math.BigInteger;
 
 @Data
 @NoArgsConstructor
@@ -19,8 +19,8 @@ import java.util.Date;
 public class Epoch {
     private long number;
     private int blockCount;
-    private int transactionCount;
-    private long totalOutput;
+    private long transactionCount;
+    private BigInteger totalOutput;
     private long startTime;
     private long endTime;
     private long maxSlot;
